started tweaking on the json data
got a response know how to pull the name out and a few other things.
need to start looking at a function that will count every number of
ojbects at the jsonobject.results[iteratethenumberhere]. Once we figure
out where to put that then we can start working on getting the random
business element squared away because that number will be the top of our
range of randomly generated number. We need to keep working on what
should be pulled out of the json as well. woot!