image carosel on page load styling
added stsyling for on page load before the button is clicked. Added
logic for assigning images to different spans and generating html. then
started working on the image carousel when you click on the image.
currecntly the logic isn't quite there yet and I can only get to the
second image...